tests/test_ocr_processor.py
"""
OCRプロセッサーのテストケース
"""

import unittest
from pathlib import Path
import tempfile
import os
from PIL import Image, ImageDraw, ImageFont
import sys

# srcディレクトリをパスに追加
sys.path.insert(0, str(Path(__file__).parent.parent / "src"))

try:
    from src.ocr_processor import OCRProcessor
except ImportError:
    # テスト実行時の代替パス
    from ocr_processor import OCRProcessor


class TestOCRProcessor(unittest.TestCase):
    """OCRプロセッサーのテストクラス"""

    def setUp(self):
        """テスト前の準備"""
        self.ocr = OCRProcessor()
        self.temp_dir = Path(tempfile.mkdtemp())

    def tearDown(self):
        """テスト後のクリーンアップ"""
        # 一時ファイルを削除
        for file in self.temp_dir.glob("*"):
            file.unlink()
        self.temp_dir.rmdir()

    def create_test_image(self, text: str, filename: str) -> Path:
        """テスト用の画像を作成"""
        image_path = self.temp_dir / filename

        # 白い背景の画像を作成
        img = Image.new("RGB", (400, 100), color="white")
        draw = ImageDraw.Draw(img)

        # テキストを描画
        try:
            # システムフォントを試行
            font = ImageFont.load_default()
        except:
            font = None

        draw.text((10, 30), text, fill="black", font=font)
        img.save(image_path)

        return image_path

    def test_supported_formats(self):
        """サポートされている画像形式のテスト"""
        supported = OCRProcessor.SUPPORTED_FORMATS
        expected_formats = {".png", ".jpg", ".jpeg", ".tiff", ".bmp", ".gif"}

        self.assertEqual(supported, expected_formats)

    def test_extract_text_simple(self):
        """シンプルなテキスト抽出のテスト"""
        try:
            # テスト画像を作成
            test_text = "Hello World"
            image_path = self.create_test_image(test_text, "test.png")

            # テキストを抽出
            result = self.ocr.extract_text(image_path)

            # 結果を検証（OCRの精度により完全一致しない場合があるため、部分一致で確認）
            self.assertIsInstance(result, str)

        except Exception as e:
            # 依存関係がない場合はスキップ
            self.skipTest(f"依存関係が不足しています: {e}")

    def test_file_not_found(self):
        """存在しないファイルのテスト"""
        non_existent_file = self.temp_dir / "non_existent.png"

        with self.assertRaises(FileNotFoundError):
            self.ocr.extract_text(non_existent_file)

    def test_unsupported_format(self):
        """サポートされていない画像形式のテスト"""
        # .txtファイルを作成
        txt_file = self.temp_dir / "test.txt"
        txt_file.write_text("This is not an image")

        with self.assertRaises(ValueError):
            self.ocr.extract_text(txt_file)

    def test_confidence_extraction(self):
        """信頼度付きテキスト抽出のテスト"""
        try:
            # テスト画像を作成
            test_text = "Test123"
            image_path = self.create_test_image(test_text, "confidence_test.png")

            # 信頼度付きでテキストを抽出
            text, confidence = self.ocr.get_text_with_confidence(image_path)

            # 結果を検証
            self.assertIsInstance(text, str)
            self.assertIsInstance(confidence, (int, float))
            self.assertGreaterEqual(confidence, 0)
            self.assertLessEqual(confidence, 100)

        except Exception as e:
            # 依存関係がない場合はスキップ
            self.skipTest(f"依存関係が不足しています: {e}")


class TestOCRIntegration(unittest.TestCase):
    """統合テストクラス"""

    def test_processor_initialization(self):
        """プロセッサーの初期化テスト"""
        # デフォルト言語での初期化
        ocr1 = OCRProcessor()
        self.assertEqual(ocr1.lang, "jpn+eng")

        # カスタム言語での初期化
        ocr2 = OCRProcessor(lang="eng")
        self.assertEqual(ocr2.lang, "eng")


if __name__ == "__main__":
    unittest.main()
