requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=9.0.0
numpy>=1.21.0
argparse
pathlib
pytest>=7.0.0